+++ camping_server1/app/controller/__init__.py
@@ -1,0 +1,3 @@
+from ..controller import detail
+from ..controller import search
+from ..controller import user
