+++ camping_server1/app/templates/index.html
@@ -11,7 +11,6 @@
     <!-- swiper.js -->
     <link rel="stylesheet" href="https://unpkg.com/swiper/swiper-bundle.min.css"/>
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" type="text/css" href="/static/css/header.css">
     <link rel="stylesheet" type="text/css" href="/static/css/index.css">
     <link rel="shorcut icon" href="../static/imgs/favicon.ico" type="image/x-icon" sizes="16x16">
 </head>
@@ -21,7 +20,7 @@
     </div>
 
     <div class="container-fluid text-white pt-5" style="width:80%;">
-        <p class="h3 ">이런 키워드는 어떠세요?</p><br>
+        <p class="h4 intro-text">이런 키워드는 어떠세요?</p><br>
         <div class="row tags">&emsp;&emsp;
             <div class="col">
                 <button type="button" class="btn btn-secondary btn-lg col-12 fw-bolder">#즐길거리</button>&nbsp;&nbsp;&nbsp;
@@ -187,36 +186,36 @@
             <div class="swiper-slide ss1">
                 <div class="row">
                     <div class="col-5">
-                        <img class="img-1" src="/static/imgs/test_img.jpg" alt="img"/>
+                        <img class="img-1" src="https://www.gocamping.or.kr/upload/camp/7038/thumb/thumb_1000_9847dZYeHK3ZjL3oNBLhykjy.jpg" alt="img"/>
                     </div>
                     <div class="col-6 promotion-1">
-                        <p class="h4 fw-bolder text-white">주목할 만한 캠핑장</p>
+                        <p class="h4 fw-bolder text-white">주변 경관이 뛰어난</p>
                         <br>
-                        <p class="text-white">캠핑장에 대한 설명</p>
+                        <p class="text-white">자연과 함께 어우러지는 사이트 간격 넓은 달빛정원글램핑</p>
                     </div>
                 </div>
             </div>
             <div class="swiper-slide ss2">
                 <div class="row">
                     <div class="col-5">
-                        <img class="img-1" src="/static/imgs/test_img2.jpg" alt="img"/>
+                        <img class="img-1" src="https://www.gocamping.or.kr/upload/camp/61/thumb/thumb_1000_7085EqJqsZa5z1eXxX7BddLq.jpg" alt="img"/>
                     </div>
                     <div class="col-6 promotion-2">
-                        <p class="h4 fw-bolder text-white">주목할 만한 캠핑장</p>
+                        <p class="h4 fw-bolder text-white">사랑하는 반려견과 함께</p>
                         <br>
-                        <p class="text-white">캠핑장에 대한 설명</p>
+                        <p class="text-white">반려견 동반이 가능한 여유있는 W글램핑</p>
                     </div>
                 </div>
             </div>
             <div class="swiper-slide ss3">
                 <div class="row">
                     <div class="col-5">
-                        <img class="img-1" src="/static/imgs/test_img3.jpg" alt="img"/>
+                        <img class="img-1" src="https://www.gocamping.or.kr/upload/camp/7227/thumb/thumb_1000_7242k4Y2QyVKFNJfIwrAYega.jpg" alt="img"/>
                     </div>
                     <div class="col-6 promotion-3">
-                        <p class="h4 fw-bolder text-white">주목할 만한 캠핑장</p>
+                        <p class="h4 fw-bolder text-white">숲의 공기와 계곡의 하모니</p>
                         <br>
-                        <p class="text-white">캠핑장에 대한 설명</p>
+                        <p class="text-white">숲과 계곡의 하모니가 어우러진 플로렌스 글램핑&캠핑</p>
                     </div>
                 </div>
             </div>
