+++ camping_server1/app/templates/layout/header.html
@@ -1,3 +1,4 @@
+<link rel="stylesheet" type="text/css" href="/static/css/header.css">
 <div class="container-fluid outer-1">
     <div class="container-fluid d-grid align-items-center" style="grid-template-columns: 0.5fr 2fr; padding-top: 10px;">
         <div class="dropdown">
@@ -68,7 +69,7 @@
 <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/typeahead.js/0.11.1/typeahead.bundle.min.js"></script>
 <script src="/static/js/tagsinput.js"></script>
-<script src="/static/js/header.css"></script>
+<script src="/static/css/header.css"></script>
 <script src="https://developers.kakao.com/sdk/js/kakao.js"></script>
 <script>
     Kakao.init('30d72d58b0b7ddc0a4ebd7fba31f8052');
